+++ R/utils.R
@@ -47,24 +47,6 @@ parse_area <- purrr::partial(parse_obj, f=parse_par)
 #' @importFrom purrr partial
 parse_page <- purrr::partial(parse_obj, f=parse_area)
 
-#' Convert bbox to magick-geometry
-#'
-#' @param x character vector with bounding boxes (top-left coordinates). Expects four integers separated by comma or space
-#'
-#' @return character vector of magick-compliant geometry string representing bounding box area in the format  "width x height +x_off +y_off" (https://www.imagemagick.org/Magick++/Geometry.html)
-#' @seealso [magick::geometry] for converting integers (four individual columns) to geometry
-#' @export
-#'
-#'
-#' @examples
-#' # area from c(0,0) to c(100,200) counting from top left
-#' bbox_to_geometry("0 0 100 200")
-#'
-bbox_to_geometry <- function(x){
-  coord_list <- lapply(strsplit(x, ",| "), as.numeric)
-  sapply(coord_list, function(x){paste0(x[3] - x[1], "x", x[4] - x[2], "+", x[1], "+", x[2])})
-}
-
 
 #' Pipe operator
 #'